Rename package promoting > promote
