+++ internal/release/promote/create_missing.go
@@ -1,4 +1,4 @@
-package promotion
+package promote
 
 import (
 	"fmt"
