+++ internal/release/promote/merge_test.go
@@ -1,4 +1,4 @@
-package promotion
+package promote
 
 import (
 	"bytes"
