+++ internal/release/promote/preview.go
@@ -1,4 +1,4 @@
-package promotion
+package promote
 
 import (
 	"fmt"
