+++ internal/release/promote/select_releases.go
@@ -1,4 +1,4 @@
-package promotion
+package promote
 
 import (
 	"bytes"
